new hp/bount/cost system v1

+++ core/src/com/betmansmall/game/gameLogic/GameField.java
@@ -777,7 +777,7 @@ public class GameField {
             shapeRenderer.setColor(Color.BLACK);
             shapeRenderer.rect(fVx + hpBarWidthIndent, fVy + currentFrameHeight - hpBarTopIndent, hpBarHPWidth, hpBarHeight);
             shapeRenderer.setColor(Color.GREEN);
-            int maxHP = creep.templateForUnit.healthPoints;
+            float maxHP = creep.templateForUnit.healthPoints;
             int hp = creep.getHp();
             hpBarHPWidth = hpBarHPWidth / maxHP * hp;
             shapeRenderer.rect(fVx + hpBarWidthIndent + hpBarSpace, fVy + currentFrameHeight - hpBarTopIndent + hpBarSpace, hpBarHPWidth - (hpBarSpace * 2), hpBarHeight - (hpBarSpace * 2));
